pyproject.toml
[project]
name = "strava"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "requests>=2.32.4",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "numpy>=2.3.0",
    "google-cloud-bigquery>=3.34.0",
    "pyarrow>=20.0.0",
    "pandas-gbq>=0.29.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/strava"]
